78595-multiple effective and future date scenario implemented AB#78595
