+++ src/UKHO.ERPFacade.API/Helpers/ScenarioBuilder.cs
@@ -22,7 +22,7 @@ namespace UKHO.ERPFacade.API.Helpers
         /// </summary>
         /// <param name="eventData"></param>
         /// <returns>List<Scenario></returns>
-        public List<Scenario> BuildScenarios(EESEvent eventData)
+        public List<Scenario> BuildScenarios(EESEventPayload eventData)
         {
             _logger.LogInformation(EventIds.IdentifyScenarioStarted.ToEventId(), "Identifying the scenarios based on received ENC content publish event.");
