+++ src/UKHO.ERPFacade.API/Models/EESEvent.cs
@@ -4,47 +4,6 @@ using System.Diagnostics.CodeAnalysis;
 namespace UKHO.ERPFacade.API.Models
 {
     [ExcludeFromCodeCoverage]
-    public class EESEvent
-    {
-        [JsonProperty("specversion")]
-        public string SpecVersion { get; set; }
-
-        [JsonProperty("type")]
-        public string Type { get; set; }
-
-        [JsonProperty("source")]
-        public string Source { get; set; }
-
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        [JsonProperty("time")]
-        public string Time { get; set; }
-
-        [JsonProperty("subject")]
-        public string Subject { get; set; }
-
-        [JsonProperty("datacontenttype")]
-        public string DataContentType { get; set; }
-
-        [JsonProperty("data")]
-        public Data Data { get; set; }
-    }
-
-    [ExcludeFromCodeCoverage]
-    public class Data
-    {
-        [JsonProperty("traceId")]
-        public string TraceId { get; set; }
-
-        [JsonProperty("products")]
-        public List<Product> Products { get; set; }
-
-        [JsonProperty("unitsOfSale")]
-        public List<UnitOfSale> UnitsOfSales { get; set; }
-    }
-
-    [ExcludeFromCodeCoverage]
     public class Product
     {
         [JsonProperty("productType")]
@@ -282,6 +241,46 @@ namespace UKHO.ERPFacade.API.Models
 
         [JsonProperty("removeProducts")]
         public List<string> RemoveProducts { get; set; }
+    }
+
+    [ExcludeFromCodeCoverage]
+    public class UnitsOfSalePrices
+    {
+        [JsonProperty("unitName")]
+        public string UnitName { get; set; }
+
+        [JsonProperty("price")]
+        public List<Price> Price { get; set; }
+    }
+
+    [ExcludeFromCodeCoverage]
+    public class Price
+    {
+        [JsonProperty("effectiveDate")]
+        public DateTime EffectiveDate { get; set; }
+
+        [JsonProperty("currency")]
+        public string Currency { get; set; }
+
+        [JsonProperty("standard")]
+        public Standard Standard { get; set; }
+    }
+
+    [ExcludeFromCodeCoverage]
+    public class Standard
+    {
+        [JsonProperty("priceDurations")]
+        public List<PriceDurations> PriceDurations { get; set; }
+    }
+
+    [ExcludeFromCodeCoverage]
+    public class PriceDurations
+    {
+        [JsonProperty("numberOfMonths")]
+        public int NumberOfMonths { get; set; }
+
+        [JsonProperty("rrp")]
+        public double Rrp { get; set; }
     }
 
     public enum Provider
